build(keycloak): Update Dockerfile for HTTPS key generation
This commit updates the Dockerfile for the Keycloak container. The changes include modifying the key generation command to specify a validity period of 36500 days for the HTTPS key. This ensures that the generated key remains valid for an extended duration. Additionally, the ownership and permissions of the HTTPS keystore file are adjusted.

- Modified the key generation command in the Dockerfile to include the "-validity 36500" flag.
- Adjusted the ownership and permissions of the HTTPS keystore file.

Author: Clay Risser <email@clayrisser.com>
Last Modified: 06-08-2023 11:26:46